Separate Orderable from Comparable
This change follows the good engineering practice of not adding 3rd
party protocol (`Comparable` is not ours!) conformance to 3rd party
types. For example, `NSData` isn't `Comparable` as of Swift 1.1 and
Swift 1.2, and a library such as Allsorts shouldn't choose to make it
conform to `Comparable` either.

Instead, only add conformance to `Orderable` and let the end users have
the freedom to do whatever else they will.