Remove the `<=>` overload of `(T?, T?)`
Since the idea is that `<=>` agrees with `<` and `==`, it's
contradictory to define `.Some(x) <=> .None == .LT` when `.Some(x)`
otherwise compares greater than `.None`.

The standard definition of the ordering between `Optional`s is
unfortunate, even if intuitive, because for lexicographically comparing
data types, there usually is a minimum (the empty sequence) but no
maximum. It would've been useful to have `Optional.None` compare the
other way, but that ship has probably sailed already, and even Haskell
has it this way, `Nothing < Just x`.

Remark: `Ended<T>` can be used when a `nil`-like value is needed to mark
the positive end cap, or "infinity" if you will.